clean up apollo client use
