+++ src/lib/takeshape/server.ts
@@ -1,9 +1,9 @@
 import { takeshapeAnonymousApiKey, takeshapeApiUrl } from '@/config';
-import { createApolloClient } from '@/lib/apollo/client';
+import { createClient } from '@/lib/apollo/client';
 import { registerApolloClient } from '@apollo/experimental-nextjs-app-support/rsc';
 
 export const { getClient: getAnonymousClient } = registerApolloClient(() => {
-  return createApolloClient({
+  return createClient({
     accessToken: takeshapeAnonymousApiKey,
     uri: takeshapeApiUrl,
     accessTokenHeader: 'Authorization',
